Add guifg colors to GitGutter
For some reason GitGutter stopped showing flags in color and also
started showing with a dark background. Suspect this has something to do
with recent changes in either Vim or tmux, but I cannot figure it out. I
update the guifg colors and this fixed the issue. Still unclear on what
caused it though.