More tweaking of prompt elements
Set background color of command execution time and status to be the same
as the default background so as not to stick out quite as much. Also
changed the status execution time message not to be red as this
indicates and error, but instead to golden to show a warning only.