Remove folds
Found that it was more difficult to get an overview of what the file
contained when so many lines were hidden/folded and not easy to glance
over. Removing the folds may mean that I need to scroll a bit more, but
hopefully it should make it easier to spot issues.